Back/requirements.txt
Flask
Flask-RESTful
SQLAlchemy
Werkzeug
PyJWT
paramiko
Flask-CORS
Flask-SQLAlchemy
